backend/requirements/base.txt
Django==6.0.3
django-ckeditor-5==0.2.20
django-jazzmin==3.0.4
django-modeltranslation==0.20.2
djangorestframework==3.17.0
pillow==12.1.1
python-decouple==3.8